+++ packages/contracts/src/systems/_NodeCreateSystem.sol
@@ -18,11 +18,11 @@ contract _NodeCreateSystem is System {
     (
       uint32 index,
       string memory nodeType,
-      uint256 roomIndex,
+      uint32 roomIndex,
       string memory name,
       string memory description,
       string memory affinity
-    ) = abi.decode(arguments, (uint32, string, uint256, string, string, string));
+    ) = abi.decode(arguments, (uint32, string, uint32, string, string, string));
     uint256 id = LibNode.getByIndex(components, index);
 
     require(id == 0, "Node: already exists");
@@ -39,7 +39,7 @@ contract _NodeCreateSystem is System {
   function executeTyped(
     uint32 index,
     string memory nodeType,
-    uint256 roomIndex,
+    uint32 roomIndex,
     string memory name,
     string memory description,
     string memory affinity
